third_party/WebKit/LayoutTests/fast/dom/DOMImplementation/missing-arguments.html

<!doctype html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
[
    "createDocument()",
    "createDocument('http://example.com/')",
    "createDocumentType()",
    "createDocumentType('svg:svg')",
    "createDocumentType('svg:svg', '-//W3C//DTD SVG 1.1//EN')"
].forEach(function(expr)
{
    shouldThrow("document.implementation." + expr);
});
</script>
</body>
</html>
